# artifacts/api-server/src/models/Review.js
import mongoose from "mongoose";
const reviewSchema = new mongoose.Schema({
    userId: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true },
    userName: { type: String, required: true },
    userAvatar: { type: String, default: null },
    menuItemId: { type: String, default: null },
    restaurantId: { type: String, default: null },
    orderId: { type: mongoose.Schema.Types.ObjectId, ref: "Order", required: true },
    rating: { type: Number, required: true, min: 1, max: 5 },
    comment: { type: String, required: true },
}, { timestamps: true });
export const Review = mongoose.model("Review", reviewSchema);