# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FoodRush - Food Delivery App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List all restaurants
      parameters:
        - name: cuisine
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: rating
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
  /restaurants/{id}/menu:
    get:
      operationId: getRestaurantMenu
      tags: [menu]
      summary: Get menu for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
  /restaurants/{restaurantId}/menu/{itemId}:
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (admin)
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                image:
                  type: string
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: menuItemId
          in: query
          schema:
            type: string
        - name: restaurantId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /users/profile:
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        role:
          type: string
        coins:
          type: number
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - coins
        - createdAt
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        cuisine:
          type: string
        image:
          type: ["string", "null"]
        rating:
          type: number
        totalRatings:
          type: number
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        minOrder:
          type: number
        address:
          type: string
        isOpen:
          type: boolean
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        menu:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"
        createdAt:
          type: string
      required:
        - id
        - name
        - cuisine
        - rating
        - totalRatings
        - deliveryTime
        - deliveryFee
        - minOrder
        - address
        - isOpen
        - featured
        - tags
        - createdAt
    MenuItem:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        restaurantName:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        image:
          type: ["string", "null"]
        rating:
          type: number
        totalRatings:
          type: number
        isVeg:
          type: boolean
        isAvailable:
          type: boolean
        isBestseller:
          type: boolean
        tags:
          type: array
          items:
            type: string
        suggestedWith:
          type: array
          items:
            type: string
      required:
        - id
        - restaurantId
        - restaurantName
        - name
        - price
        - category
        - rating
        - totalRatings
        - isVeg
        - isAvailable
        - isBestseller
        - tags
        - suggestedWith
...
    Review:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        menuItemId:
          type: ["string", "null"]
        restaurantId:
          type: ["string", "null"]
        orderId:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - orderId
        - rating
        - comment
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        menuItemId:
          type: string
        restaurantId:
          type: string
        orderId:
          type: string
        rating:
          type: number
        comment:
          type: string
      required:
        - orderId
        - rating
        - comment
...
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        avatar:
          type: string
...